- Updated Simple Theme to do a bit better highlighting of XML - Updated the Terminal script to some version that Allan suggested on IRC a couple of days ago
- Added $0 after the XML tag snippets.

Can snippet editors please do this for *every* snippet they define, unless they have a very good reason not to? It sucks to be able to smart tab to different places of the snippet, but end up caught in the snippet and you have to arrow your way out of the snippet.

git-svn-id: http://svn.textmate.org/trunk/Bundles/XML.tmbundle@850 dfb7d73b-c2ec-0310-8fea-fb051d288c6d