Update XML.plist
adding a filetype for a vendor specific XML file. This is for BigFix content XML files.

There are examples here: https://github.com/jgstew/bigfix-content/tree/master/fixlet

They are currently renamed to end with ".xml" in order to have proper syntax highlighting, but ideally they would have the proper syntax highlighting with the ".bes" extension directly since that is how they are used natively. 

I'm not sure if vendor specific filetypes are appropriate or not, or what the alternative is.